HW6 Completed

+++ back/ecommerce-blockbeam22y/src/modules/products/products.service.ts
@@ -1,9 +1,13 @@
-import { Inject, Injectable } from '@nestjs/common';
+import {
+  BadRequestException,
+  ConflictException,
+  Injectable,
+  NotFoundException,
+} from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import Product from './products.entity';
 import Category from '../categories/categories.entity';
 import { Repository } from 'typeorm';
-import { PreloadProduct } from 'src/utils/preloadData';
 
 @Injectable()
 export class ProductsService {
@@ -12,7 +16,6 @@ export class ProductsService {
     private readonly productsRepository: Repository<Product>,
     @InjectRepository(Category)
     private readonly categoriesRepository: Repository<Category>,
-    @Inject('preload') private readonly preloadData: PreloadProduct[],
   ) {}
 
   async getProducts(page: number, limit: number) {
@@ -26,18 +29,34 @@ export class ProductsService {
   }
 
   async getProductById(id: string) {
-    return this.productsRepository.findOne({
+    const product = await this.productsRepository.findOne({
       where: { id },
       relations: {
         category: true,
       },
     });
+
+    if (product) {
+      return product;
+    } else {
+      throw new NotFoundException(`Couldn't find product with id '${id}'`);
+    }
   }
 
   async createProduct({ categoryId, ...productData }) {
+    const category = await this.categoriesRepository.findOneBy({
+      id: categoryId,
+    });
+
+    if (!category) {
+      throw new BadRequestException(
+        `Couldn't find category with id '${categoryId}'`,
+      );
+    }
+
     const product = this.productsRepository.create({
       ...productData,
-      category: { id: categoryId },
+      category,
     });
     await this.productsRepository.save(product);
 
@@ -45,14 +64,35 @@ export class ProductsService {
   }
 
   async updateProduct(id: string, productData: Partial<Product>) {
-    await this.productsRepository.update(id, productData);
+    const product = await this.productsRepository.findOneBy({ id });
 
-    return id;
+    if (product) {
+      await this.productsRepository.update(id, productData);
+
+      return product.id;
+    } else {
+      throw new NotFoundException(`Couldn't find product with id '${id}'`);
+    }
   }
 
   async deleteProduct(id: string) {
-    await this.productsRepository.delete(id);
+    const product = await this.productsRepository.findOne({
+      where: { id },
+      relations: {
+        orderDetails: true,
+      },
+    });
 
-    return id;
+    if (!product) {
+      throw new NotFoundException(`Couldn't find product with id '${id}'`);
+    } else if (product.orderDetails.length) {
+      throw new ConflictException(
+        `Product with id '${id}' is already included in an order`,
+      );
+    } else {
+      await this.productsRepository.delete(id);
+
+      return product.id;
+    }
   }
 }
