+++ back/ecommerce-blockbeam22y/src/modules/users/users.service.ts
@@ -1,37 +1,37 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import User from './users.entity';
+import Order from '../orders/entities/orders.entity';
 import { Repository } from 'typeorm';
 
 @Injectable()
 export class UsersService {
   constructor(
     @InjectRepository(User) private readonly usersRepository: Repository<User>,
+    @InjectRepository(Order)
+    private readonly ordersRepository: Repository<Order>,
   ) {}
 
   async getUsers(page: number, limit: number) {
-    const users = await this.usersRepository.find({
+    return this.usersRepository.find({
       take: limit,
       skip: (page - 1) * limit,
     });
-
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    return users.map(({ password, ...user }) => user);
   }
 
   async getUserById(id: string) {
-    // eslint-disable-next-line @typescript-eslint/no-unused-vars
-    const { password, orders, ...user } = await this.usersRepository.findOne({
+    const user = await this.usersRepository.findOne({
       where: { id },
       relations: {
         orders: true,
       },
     });
 
-    return {
-      ...user,
-      orders: orders.map(({ id, date }) => ({ id, date })),
-    };
+    if (user) {
+      return user;
+    } else {
+      throw new NotFoundException(`Couldn't find user with id '${id}'`);
+    }
   }
 
   async createUser(userData: Omit<User, 'id' | 'orders'>) {
@@ -42,14 +42,34 @@ export class UsersService {
   }
 
   async updateUser(id: string, userData: Partial<User>) {
-    await this.usersRepository.update(id, userData);
+    const user = await this.usersRepository.findOneBy({ id });
 
-    return id;
+    if (user) {
+      await this.usersRepository.update(id, userData);
+
+      return user.id;
+    } else {
+      throw new NotFoundException(`Couldn't find user with id '${id}'`);
+    }
   }
 
   async deleteUser(id: string) {
-    await this.usersRepository.delete(id);
+    const user = await this.usersRepository.findOne({
+      where: { id },
+      relations: {
+        orders: true,
+      },
+    });
 
-    return id;
+    if (user) {
+      for await (const order of user.orders) {
+        await this.ordersRepository.delete(order);
+      }
+
+      await this.usersRepository.delete(id);
+      return user.id;
+    } else {
+      throw new NotFoundException(`Couldn't find user with id '${id}'`);
+    }
   }
 }
